misc: random string creator
It is pretty useful to count with a function that allows us generating random strings.

In Fact this matter is normally applied through the technique of creation of files inside /tmp. Such approach are going to avoid any possible race condition among several instances of a program utilizing temporal files for any task that they carry out.

If the size of random string has not been determined when using the function, default value shall be set as 8 "like the ancient DOS files".